requirements.txt
google-generativeai
argparse